Refactor the entire project
There are many changes in this commit. It's almost a new code but with
the same game. I broke a lot of stuff, particularly the collision
algorithm which is not working now.

I removed a bunch of non needed static variables and functions,
removed any license header to reduce file verbosity. There are new
functions wirthin the vector classes, improved boolean checks,
improved logging and remove a bunch of dead code.

Signed-off-by: Rafael Campos Nunes <rcamposnunes@outlook.com>